third_party/blink/web_tests/fullscreen/full-screen-request-expected.txt

EVENT(webkitfullscreenchange) TEST(document.webkitIsFullScreen==true) OK
END OF TEST
